Refactor and clean up JSON parser
