+++ common/src/java/org/apache/hadoop/hive/conf/HiveConf.java
@@ -486,8 +486,9 @@ public class HiveConf extends Configuration {
                     "Only in effect when the cache is enabled"),
     MSC_CACHE_RECORD_STATS("hive.metastore.client.cache.v2.recordStats", false,
             "This property enables recording metastore client cache stats in DEBUG logs"),
-    QUERY_PLAN_CACHE_ENABLED("hive.query.plan.cache.enabled", true,
-        "This property enables Query Plan Cache"),
+    CBO_PLAN_SERIALIZATION_DESERIALIZATION_ENABLED("hive.cbo.plan.serialization.deserialization.enabled",
+        true, "This property enables serialization and deserialization for cbo plans. Currently" +
+        "for internal use only."),
     // QL execution stuff
     DFS_XATTR_ONLY_SUPPORTED_ON_RESERVED_NAMESPACE("dfs.xattr.supported.only.on.reserved.namespace", false,
       "DFS supports xattr only on Reserved Name space (/.reserved/raw)"),
